Bumping object loader source code and ignoring most of this project
This is a test / emergency commit to try to get the latest objloader
source code into my other project (charimsa). I've pointed the entry
file to source since there's no reason for a build step. Copied the
latest objloader source from three
https://github.com/mrdoob/three.js/blob/master/examples/js/loaders/OBJLoader.js
in an attempt to debug why a multimaterial on a buffergeometry isn't
showing up